requirements.txt
Flask>=2.0.1
gunicorn>=20.1.0
Pillow>=8.3.1
pdf2image>=1.16.0
pyzbar>=0.1.8
werkzeug>=2.0.1